+++ resources/js/addon.js
@@ -1,4 +1,6 @@
-import '../css/addon.scss';
+/* @deprecated: Alias the Statamic v4 scales so the same classes can work in v3 as well, drop when v3 is dropped */
+import '../css/addon.css';
+import '../css/jstw.css';
 
 import Provider from './provider'