+++ resources/js/components/AttributesPanel.vue
@@ -3,15 +3,15 @@
     <div class="bts-panel">
         <div v-if="items.length">
             <div v-for="(item, i) in items">
-                <div class="font-bold px-2 bts-py-1.25 bg-grey-10 title-case border-b flex items-center cursor-pointer" @click="activeItem = i">
-                    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-4 h-4 bts-arrow text-grey-70" :class="{ 'rotate-90': activeItem === i }">
+                <div class="font-bold jstw-px-4 jstw-py-2.5 jstw-bg-gray-100 title-case border-b flex items-center cursor-pointer" @click="activeItem = i">
+                    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-4 h-4 bts-arrow jstw-text-gray-700" :class="{ 'rotate-90': activeItem === i }">
                         <path stroke-linecap="round" stroke-linejoin="round" d="M8.25 4.5l7.5 7.5-7.5 7.5" />
                     </svg>
                     {{ titles[item.type] }}
                 </div>
-                <div class="p-2 border-b space-y-1.5" v-if="activeItem === i">
-                    <div v-for="(field, name) in fields(item.type)">
-                        <label v-if="field.type === 'toggle'" class="flex items-baseline bts-gap-0.5 font-normal">
+                <div class="jstw-p-4 jstw-pt-1 border-b" v-if="activeItem === i">
+                    <div v-for="(field, name) in fields(item.type)" class="jstw-mt-3">
+                        <label v-if="field.type === 'toggle'" class="flex items-baseline jstw-gap-1 font-normal">
                             <input
                                 type="checkbox"
                                 v-model="item.attrs[name]"
@@ -25,13 +25,13 @@
                             <input
                                 type="text"
                                 v-model="item.attrs[name]"
-                                class="input h-8 p-1 border text-sm border-grey-50 rounded mt-1"
+                                class="input h-8 jstw-p-2 border text-sm jstw-border-gray-500 rounded jstw-mt-2"
                             />
                         </label>
                     </div>                    
                 </div>
             </div>
-            <div class="flex items-center justify-end space-x-1 font-normal px-2 py-1.5">
+            <div class="flex items-center justify-end font-normal jstw-px-4 jstw-py-3">
                 <button
                     @click="apply"
                     class="btn btn-sm">
@@ -39,7 +39,7 @@
                 </button>
             </div>
         </div>
-        <div v-else class="p-4 w-40 text-center text-grey-40">
+        <div v-else class="jstw-p-8 w-40 text-center jstw-text-gray-400">
             <svg viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg" class="w-12 h-12">
                 <path d="M23.168.832.809 23.191M21 12c0 4.937-4.063 9-9 9s-9-4.063-9-9 4.063-9 9-9 9 4.063 9 9Z" style="fill:none;fill-rule:nonzero;stroke-width:1.5px" stroke="currentColor" />
             </svg>